# swagger/openApi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "version": "1.0.0",
        "description": "API for managing products"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/api": {
            "get": {
                "summary": "Api health endpoint",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "json"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "requestBody": {
                    "description": "The product object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/products/{productId}": {
            "get": {
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "summary": "Update a product by ID",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The product object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a product by ID",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "The ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components":{
        "schemas":{
            "Product": {
                "type": "object",
                "properties": {
                    "productId":{
                        "type": "integer"
                    },
                    "productName":{
                        "type": "string"
                    },
                    "productOwnerName":{
                        "type": "string"
                    },
                    "Developers": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }        
                    },
                    "scrumMasterName":{
                        "type": "string"
                    },
                    "startDate":{
                        "type": "string"
                    },
                    "methodology":{
                        "type": "string"
                    }
                }
            }
        }
    }    
}